refactor(bot): extract provisional-strategy task loops into StrategyTaskMixin (phase 5)
Second cut of the AuramaurBot split. The periodic task loops that construct and
drive each provisional pillar (bias_harvest, entailment_arb, cross_venue_arb,
econ_indicator, intraday_drift, hydro_watch, weather_temp, resolution_lens,
oddlot_tender) move verbatim into auramaur/bot_strategy_tasks.py as
StrategyTaskMixin. Pure structural move — pillar classes stay lazily imported
inside each loop, the loops still operate on the bot's self, behavior identical.

bot.py: 3462 -> 3232 lines (587 total across both phase-5 cuts so far). The
money-critical arb-execution block is left for a focused follow-up. Full suite:
837 passed.
