+++ auramaur/bot.py
@@ -52,11 +52,12 @@ from auramaur.strategy.news_reactor import NewsReactor
 from auramaur.strategy.resolution_tracker import ResolutionTracker
 from auramaur.strategy.technical import TechnicalAnalyzer
 from auramaur.bot_exits import ExitExecutionMixin
+from auramaur.bot_strategy_tasks import StrategyTaskMixin
 
 log = structlog.get_logger()
 
 
-class AuramaurBot(ExitExecutionMixin):
+class AuramaurBot(ExitExecutionMixin, StrategyTaskMixin):
     """Main bot orchestrator running concurrent async tasks."""
 
     def __init__(
@@ -2475,237 +2476,6 @@ class AuramaurBot(ExitExecutionMixin):
                 show_error(f"News reactor failed: {e}")
             interval = self.settings.hybrid.news_cycle_seconds if self._hybrid else 60
             await asyncio.sleep(interval)
-
-    async def _task_bias_harvest(self) -> None:
-        """Periodic favorite-longshot bias harvest scan (paper-forced by config)."""
-        from auramaur.strategy.bias_harvest import BiasHarvestPillar
-
-        pillar = BiasHarvestPillar(
-            db=self._components["db"],
-            settings=self.settings,
-            discovery=self._components["discovery"],
-            exchange=self._components["exchange"],
-            risk_manager=self._components["risk_manager"],
-            pnl_tracker=self._components["pnl_tracker"],
-            calibration=self._components["calibration"],
-        )
-        interval = max(60, self.settings.bias_harvest.interval_seconds)
-        while self._running:
-            if await self._check_kill_switch():
-                return
-            try:
-                await pillar.run_once()
-            except Exception as e:
-                log.error("bias_harvest.cycle_error", error=str(e))
-            await asyncio.sleep(interval)
-
-    async def _task_entailment_arb(self) -> None:
-        """Periodic entailment-arbitrage scan (paper-forced by config)."""
-        from auramaur.strategy.entailment_arb import EntailmentArbPillar
-
-        pillar = EntailmentArbPillar(
-            db=self._components["db"],
-            settings=self.settings,
-            discovery=self._components["discovery"],
-            exchange=(self._components.get("exchanges") or {}).get("polymarket"),
-            risk_manager=self._components["risk_manager"],
-            pnl_tracker=self._components["pnl_tracker"],
-            analyzer=self._components.get("analyzer"),
-            # Kalshi econ-bin ladder arb fetches its own series; pass the Kalshi
-            # discovery if present (no-op when Kalshi isn't configured).
-            kalshi_discovery=(self._components.get("discoveries") or {}).get("kalshi"),
-            exchanges=self._components.get("exchanges") or {},
-        )
-        interval = max(60, self.settings.entailment_arb.interval_seconds)
-        while self._running:
-            if await self._check_kill_switch():
-                return
-            try:
-                await pillar.run_once()
-            except Exception as e:
-                log.error("entailment.cycle_error", error=str(e))
-            await asyncio.sleep(interval)
-
-    async def _task_cross_venue_arb(self) -> None:
-        """Periodic cross-venue (Poly×Kalshi) semantic-equivalence arb scan.
-        Paper-forced by config; no-op unless Kalshi discovery is wired."""
-        from auramaur.strategy.cross_venue_arb import CrossVenueArbPillar
-
-        pillar = CrossVenueArbPillar(
-            db=self._components["db"],
-            settings=self.settings,
-            discovery=(self._components.get("discoveries") or {}).get("polymarket"),
-            exchange=(self._components.get("exchanges") or {}).get("polymarket"),
-            risk_manager=self._components["risk_manager"],
-            pnl_tracker=self._components["pnl_tracker"],
-            analyzer=self._components.get("analyzer"),
-            kalshi_discovery=(self._components.get("discoveries") or {}).get("kalshi"),
-            exchanges=self._components.get("exchanges") or {},
-        )
-        interval = max(60, self.settings.cross_venue_arb.interval_seconds)
-        while self._running:
-            if await self._check_kill_switch():
-                return
-            try:
-                await pillar.run_once()
-            except Exception as e:
-                log.error("cross_venue.cycle_error", error=str(e))
-            await asyncio.sleep(interval)
-
-    async def _task_econ_indicator(self) -> None:
-        """Periodic data-driven Kalshi econ-indicator bin pricing (paper-forced)."""
-        from auramaur.data_sources.fred import FREDSource
-        from auramaur.strategy.econ_indicator import EconIndicatorPillar
-
-        kalshi_discovery = (self._components.get("discoveries") or {}).get("kalshi")
-        kalshi_exchange = (self._components.get("exchanges") or {}).get("kalshi")
-        if kalshi_discovery is None or kalshi_exchange is None or not self.settings.fred_api_key:
-            log.info("econ_indicator.disabled", reason="missing kalshi or FRED key")
-            return
-        pillar = EconIndicatorPillar(
-            db=self._components["db"],
-            settings=self.settings,
-            kalshi_discovery=kalshi_discovery,
-            fred_source=FREDSource(api_key=self.settings.fred_api_key),
-            exchange=kalshi_exchange,
-            risk_manager=self._components["risk_manager"],
-            pnl_tracker=self._components["pnl_tracker"],
-            calibration=self._components["calibration"],
-        )
-        interval = max(60, self.settings.econ_indicator.interval_seconds)
-        while self._running:
-            if await self._check_kill_switch():
-                return
-            try:
-                await pillar.run_once()
-            except Exception as e:
-                log.error("econ_indicator.cycle_error", error=str(e))
-            await asyncio.sleep(interval)
-
-    async def _task_intraday_drift(self) -> None:
-        """Measurement spike: track post-signal price drift toward the LLM estimate
-        (no trading). Gates the intraday-convergence strategy on real evidence."""
-        from auramaur.monitoring.intraday_drift import IntradayDriftTracker
-
-        tracker = IntradayDriftTracker(
-            db=self._components["db"],
-            settings=self.settings,
-            discovery=self._components["discovery"],
-        )
-        interval = max(60, self.settings.intraday_drift.interval_seconds)
-        while self._running:
-            if await self._check_kill_switch():
-                return
-            try:
-                await tracker.run_once()
-            except Exception as e:
-                log.error("intraday_drift.cycle_error", error=str(e))
-            await asyncio.sleep(interval)
-
-    async def _task_hydro_watch(self) -> None:
-        """Alert when a tradeable hydrology market appears (compHydro moat armed)."""
-        from auramaur.monitoring.hydro_market_watch import HydroMarketWatcher
-
-        watcher = HydroMarketWatcher(
-            db=self._components["db"],
-            settings=self.settings,
-            discoveries=self._components.get("discoveries") or {},
-            alerts=self._components.get("alerts"),
-        )
-        interval = max(300, self.settings.hydro_watch.interval_seconds)
-        while self._running:
-            if await self._check_kill_switch():
-                return
-            try:
-                await watcher.run_once()
-            except Exception as e:
-                log.error("hydro_watch.cycle_error", error=str(e))
-            await asyncio.sleep(interval)
-
-    async def _task_weather_temp(self) -> None:
-        """Open-Meteo ensemble pricing of Polymarket city-temperature bins (paper)."""
-        from auramaur.data_sources.openmeteo import OpenMeteoSource
-        from auramaur.strategy.weather_temp import WeatherTempPillar
-
-        pillar = WeatherTempPillar(
-            db=self._components["db"],
-            settings=self.settings,
-            discovery=self._components["discovery"],
-            exchange=self._components["exchange"],
-            risk_manager=self._components["risk_manager"],
-            pnl_tracker=self._components["pnl_tracker"],
-            calibration=self._components["calibration"],
-            weather=OpenMeteoSource(),
-        )
-        interval = max(60, self.settings.weather_temp.interval_seconds)
-        while self._running:
-            if await self._check_kill_switch():
-                return
-            try:
-                await pillar.run_once()
-            except Exception as e:
-                log.error("weather_temp.cycle_error", error=str(e))
-            await asyncio.sleep(interval)
-
-    async def _task_resolution_lens(self) -> None:
-        """Periodic resolution-language lens scan (paper-forced by config)."""
-        from auramaur.strategy.resolution_lens import ResolutionLensPillar
-
-        pillar = ResolutionLensPillar(
-            db=self._components["db"],
-            settings=self.settings,
-            discovery=self._components["discovery"],
-            exchange=self._components["exchange"],
-            risk_manager=self._components["risk_manager"],
-            pnl_tracker=self._components["pnl_tracker"],
-            calibration=self._components["calibration"],
-            analyzer=self._components.get("analyzer"),
-            # Phase 3: evidence-grounded comprehension taps the shared aggregator.
-            aggregator=self._components.get("aggregator"),
-        )
-        interval = max(60, self.settings.resolution_lens.interval_seconds)
-        while self._running:
-            if await self._check_kill_switch():
-                return
-            try:
-                await pillar.run_once()
-            except Exception as e:
-                log.error("lens.cycle_error", error=str(e))
-            await asyncio.sleep(interval)
-
-    async def _task_oddlot_tender(self) -> None:
-        """EDGAR odd-lot tender scan (detection always; entries paper-forced)."""
-        from auramaur.data_sources.edgar import EdgarClient
-        from auramaur.strategy.oddlot_tender import OddLotTenderPillar
-
-        edgar = EdgarClient()
-        equity = None
-        if self.settings.ibkr.enabled:
-            from auramaur.exchange.ibkr_equity import IBKREquityClient
-            equity = IBKREquityClient(self.settings)
-        pillar = OddLotTenderPillar(
-            db=self._components["db"],
-            settings=self.settings,
-            edgar=edgar,
-            analyzer=self._components.get("analyzer"),
-            alerts=self._components.get("alerts"),
-            equity_client=equity,
-            pnl_tracker=self._components["pnl_tracker"],
-        )
-        interval = max(600, self.settings.oddlot_tender.interval_seconds)
-        try:
-            while self._running:
-                if await self._check_kill_switch():
-                    return
-                try:
-                    await pillar.run_once()
-                except Exception as e:
-                    log.error("oddlot.cycle_error", error=str(e))
-                await asyncio.sleep(interval)
-        finally:
-            await edgar.close()
-            if equity is not None:
-                await equity.close()
 
     async def _task_build_guard(self) -> None:
         """Warn when the on-disk checkout moves past the running process.
